opsctl/requirements.txt
typer==0.15.1
pyyaml==6.0.1
requests==2.25.0